pyproject.toml
[project]
name = "sesc-inf-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.2.1",
]
